requirements.txt
Django
django-cors-headers
djangorestframework-simplejwt
djangorestframework
django-filter~=23.2
django-secured-fields==0.4.2
django-filter==23.2
psycopg2-binary
python-dotenv
redis
daphne
django-redis
channels~=4.0.0
channels-redis
celery>=5.0
uvicorn
uvicorn[standard]
gunicorn
Pillow
faker
